+++ src/app/core/layouts/footer/footer.component.ts
