+++ src/app/core/models/event.model.ts
