+++ src/app/core/services/event.service.ts
@@ -1,7 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Event } from '../models/Event';
+import { Event } from '../models/event.model';
+
 
 @Injectable({
   providedIn: 'root'
