+++ src/app/features/event-acquisition/components/ticket-selection/ticket-selection.component.html
@@ -1,13 +1,13 @@
-<section class="flex justify-content-center align-items-center w-full h-full border-round-lg bg-black-alpha-90 ">
+<section class="flex justify-content-center align-items-center w-full h-full border-round-lg bg-primary ">
     <div class="w-full h-max px-5 ">
         <p-divider />
-        <div class="flex justify-content-around align-items-center w-full">       
-            <div class="flex justify-content-center align-items-center w-6 gap-3">
+        <div class="flex flex-column md:flex-row justify-content-around align-items-center w-full">       
+            <div class="flex justify-content-center align-items-center w-6 gap-3 mb-3 md:mb-0">
                 <div class="flex flex-column p-1">
-                    <p class="text-4xl text-white font-bold w-max h-min mb-2 mt-0">
+                    <p class=" text-2xl sm:text-4xl text-secondary font-bold w-max h-min mb-2 mt-0">
                         {{event.title}}
                     </p>
-                    <span class="text-white">
+                    <span class="text-secondary">
                         {{event.description}}
                     </span>
                 </div>
