+++ src/app/features/event-acquisition/components/ticket-selection/ticket-selection.component.scss
