+++ src/app/features/event-acquisition/pages/event-acquisition.component.scss
