+++ src/app/shared/components/language-selector/language-selector.component.scss
